requirements.txt
pytest==8.0.0
selenium==4.20.0
requests==2.31.0
SQLAlchemy==1.4.50
psycopg2-binary==2.9.9
greenlet==3.0.3
typing_extensions==4.11.0
allure-pytest==2.13.5